+++ arcile-app/app/src/main/java/dev/qtremors/arcile/feature/storagecleaner/StorageCleanerViewModel.kt
@@ -3,7 +3,7 @@ package dev.qtremors.arcile.feature.storagecleaner
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
-import dev.qtremors.arcile.core.storage.data.StorageCleanerScanner
+import dev.qtremors.arcile.core.storage.domain.StorageCleanerScanner
 import dev.qtremors.arcile.core.storage.domain.CleanerGroup
 import dev.qtremors.arcile.core.storage.domain.CleanerGroupType
 import dev.qtremors.arcile.core.storage.domain.FileRepository