+++ arcile-app/app/src/test/java/dev/qtremors/arcile/feature/browser/delegate/NavigationDelegateTest.kt
@@ -1,7 +1,7 @@
 package dev.qtremors.arcile.feature.browser.delegate
 
 import androidx.lifecycle.SavedStateHandle
-import dev.qtremors.arcile.core.storage.data.BrowserPreferencesStore
+import dev.qtremors.arcile.core.storage.domain.BrowserPreferencesStore
 import dev.qtremors.arcile.core.storage.domain.FileModel
 import dev.qtremors.arcile.core.storage.domain.FileRepository
 import dev.qtremors.arcile.core.storage.domain.SearchFilters