+++ arcile-app/app/src/test/java/dev/qtremors/arcile/testutil/FakeBrowserPreferencesStore.kt
@@ -1,6 +1,6 @@
 package dev.qtremors.arcile.testutil
 
-import dev.qtremors.arcile.core.storage.data.BrowserPreferencesStore
+import dev.qtremors.arcile.core.storage.domain.BrowserPreferencesStore
 import dev.qtremors.arcile.core.storage.domain.BrowserPreferences
 import dev.qtremors.arcile.core.storage.domain.BrowserPresentationPreferences
 import kotlinx.coroutines.flow.Flow